changed the way U is set in ABE.
Switched the way default values are selected for stands in ABE. 
First, the value of the "U" column in the CSV file is used; if not set, then the default U of the assigned STP is used (if a STP is assigned in the data file). Else the U value of the unit is assigned (which is the U of the default STP from the managing agent). In addition to that, U can be set programmatically via Javascript.