Last commit with Subversion!
iLand used subversion as code repository since early 2009 - the first commit was on the 12th of Feb 2009, so 15 years ago!
During the Vienna years, the SVN repository was located at on a virtual machine provided by BOKU Vienna (the webserver hosting iland.boku.ac.at). After moving to TU Munich, the iLand website changed to iland-model.org, and lives since then on a virtual machined provided by LRZ / TUM. 
Now it is time to move on - the iLand source code will move to GIT and GitHub - I hope this will make it more accessible to the community!