Optimized way of loading text files
The old method loaded the whole text file into a string. This seems to
be limited to text files < 3.5 GB (roughly).
The additional new method reads text file line by line and returns a
QStringList. This is now used for loading CSVFile and initalization.
The new method can digest larger file, and is faster.