Fixing issues with large simulations
We saw several crashes in a simulation study with very large management
operations (mimicking disturbances). I looked at some of the potential
culprits (and added some optimizations there). The big one is the
"filter()" function of management.cpp: this used a loop and erase() to
do the filtering. This scales very badly for very large lists (say, 20
million trees), as QList is *actually* an array underneath (and the old
implementation assumed a linked list).
In any case, the difference is *huge* for large lists.
before: 40 *minutes*
after: 3 *seconds*

The implementation is now in management::filter(), but should be rolled
out to other places where the same pattern appears.