Added option for manual exit of activities
This is useful, when an activity triggers other activites (e.g. via
signals), and is this thus the "active" activity. For example, a strip-
cut activity might only trigger other events, but should be active until
all strips are processed.

To switch on:
onCreate: function() { this.manualExit = true; }

Currently the an activity can be forced to end with:
this.enabled = false;