Addition of signal parameters
Sending of signals in ABE world now includes the option to add a signal
parameter (any object).

To use the signal parameters we have now a new stand property
`signalParameter` that is set to the signal value. In case of Javascript
functions, the parameter is directly passed to the function.