Update of Singletons
Modernized how singletons are used in the model. The original code (2009
or so) used a explicit if()  for the instance() function.
As I realized now in modern (C++ 11) an alternative approach with local
static variables is available.

I think iLand got a bit faster, but would need some more stringent tests
to really proof that.