base version of understorey modules
this is a skeleton of the Understorey module for iLand. Far from
finished, but some basics are laid down.

I think the decision to also use modern c++ in iLand makes really sense,
even if there are "old" and new parts in the codebase. But it is
definitely nice to work with that :)

Many changes are due to changes not directly related to understorey:

* NSAPLINGS is not a c++ constexpr
* the GlobalSettings instance uses now a more modern singleton approach
(which is hopefully a little bit faster; I tested but if there is an
improvement it is probably low)

Understorey:

basic structure for submodule; somewhat similar to microclimate; the
main container for the actual understorey plant is per RU, which allows
efficient parallelization on RU basis.
All understorey related data is stored in a single Understorey object
(that holds a vector of RU-level objects which hold a vector of cells
which hold an array of plants). It is all a single blob in memory;
currently 30kB per RU.