fixed a annoying warning
For ages, there were a couple of warnings such as "libpng warning: iCCP:
known incorrect sRGB profile". Quite annoying.

A had a deeper look now, and found a solution: by using a breakpoint
when the warning is logged I was able to find the files that cause it. I
saved now the PNG without the color profile.