Fixes in ABE
* chains of signals: code emitted from signals can now again trigger
signals (and so forth). Now the iteration depth is unlimited (there is
an infinite loop detection, though)
* when an activity is executed from a signal it typically does not
affect the time line of activities; however, there is now an exception
to that rule: if the triggered activity is a finalHarvest, then it makes
sense that the time line of the stand is altered.