Fix in encoding of direction of external seeds
Thanks to the students of the "waldsimulation" class in 2025! They found
it is not working when direction is set to "n" (north).
I had a look and the code was weird - in the end it was an error in
counting bits and to make it work for three directions (s,e,w) the code
made not much sense, mixing X and Y directions :(
Now it works as expected.