Working on dead trees
Fixed two or three crashes, which in the end were caused by a logical
error.
Improved visualization of snags; now you can click on a snag and see
details in the detail-view. Some improvements there too, clicking on the
visualization now shows the right details or no details (and not random
tree info)