restructure execution of regen, revamp of timers
As a prep for understory, establishment and sapling growth is not
packaged differently; now a single function runs all processes related
to regen per RU (which allows temporary data for the regen processes).

In addition, ticktack.cpp was removed, and the timing code uses now
QElapsedTimer on all platforms. Added a new timing overview to the log.